+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Change the schema of MQTT and SMTP credentials to make them same than those of DBMS (todo #158)
     - Rename schemas according to their directory
     - Define new 'wantsAccount' and 'wantsPassword' for MQTT and SMTP gateways
+    - Fix EP bootstrapping when main log file is not defined
 
 
 ### 4.14.1